web/requirements.txt
django==3.1.4
djangorestframework==3.12.2
environs==7.3.0
psycopg2-binary==2.8.6
social-auth-app-django==4.0.0
social-auth-core==3.3.3
uwsgi==2.0.19.1
